Added the Video Adding and uploaded the functionality.
Show the progress indicator while uploading the Video.
Also show the Snack Bar(toast) after successfully Uploaded the Video on the server.

Also Display the Link of the Uploaded Video Url, So that the user know the Video is uploaded.

Also mark the Video link read only so that no one copy the Video URL link directly.

Edit The lecture page in which the Edit page gets the Current lecture name and if the user change this title into the new title and click on the save button the old title changes into the new title.